@@ -15,22 +15,22 @@
     "spartacus"
   ],
   "peerDependencies": {
-    "@angular/common": "~10.2.4",
-    "@angular/compiler": "~10.2.4",
-    "@angular/core": "~10.2.4",
-    "@angular/forms": "~10.2.4",
-    "@angular/localize": "~10.2.4",
-    "@angular/platform-browser": "~10.2.4",
-    "@angular/router": "~10.2.4",
-    "@angular/service-worker": "~10.2.4",
+    "@angular/common": "~12.0.5",
+    "@angular/compiler": "~12.0.5",
+    "@angular/core": "~12.0.5",
+    "@angular/forms": "~12.0.5",
+    "@angular/localize": "~12.0.5",
+    "@angular/platform-browser": "~12.0.5",
+    "@angular/router": "~12.0.5",
+    "@angular/service-worker": "~12.0.5",
     "blob-util": "^2.0.2",
     "file-saver": "^2.0.2",
-    "@ngrx/effects": "~10.0.0",
-    "@ngrx/store": "~10.0.0",
-    "@spartacus/core": "^3.4.0",
-    "@spartacus/storefront": "^3.4.0",
+    "@ngrx/effects": "~12.1.0",
+    "@ngrx/store": "~12.1.0",
+    "@spartacus/core": "^4.3.0",
+    "@spartacus/storefront": "^4.2.0",
     "bootstrap": "4.5.0",
-    "rxjs": "^6.5.4",
-    "zone.js": "^0.10.2"
+    "rxjs": "^6.6.0",
+    "zone.js": "~0.11.4"
   }
 }
