@@ -15,27 +15,41 @@
     "storefront"
   ],
   "peerDependencies": {
-    "@angular/common": "~10.2.4",
-    "@angular/compiler": "~10.2.4",
-    "@angular/core": "~10.2.4",
-    "@angular/forms": "~10.2.4",
-    "@angular/localize": "~10.2.4",
-    "@angular/platform-browser": "~10.2.4",
-    "@angular/router": "~10.2.4",
-    "@angular/service-worker": "~10.2.4",
-    "@ng-bootstrap/ng-bootstrap": "^6.0.0",
-    "@ng-select/ng-select": "^5.0.0",
-    "@ngrx/effects": "~10.0.0",
-    "@ngrx/store": "~10.0.0",
-    "@spartacus/core": "3.4.0",
-    "@spartacus/user": "3.4.0",
-    "@syncpilot/bpool-guest-lib": "^0.2.6",
+    "@angular/common": "~12.0.5",
+    "@angular/compiler": "~12.0.5",
+    "@angular/core": "~12.0.5",
+    "@angular/forms": "~12.0.5",
+    "@angular/localize": "~12.0.5",
+    "@angular/platform-browser": "~12.0.5",
+    "@angular/router": "~12.0.5",
+    "@angular/service-worker": "~12.0.5",
+    "@ng-bootstrap/ng-bootstrap": "~10.0.0",
+    "@ng-select/ng-select": "~7.0.1",
+    "@ngrx/effects": "~12.1.0",
+    "@ngrx/store": "~12.1.0",
+    "@spartacus/core": "4.3.0",
+    "@spartacus/user": "4.2.1",
+    "@spartacus/styles": "4.2.1",
+    "@spartacus/storefront": "4.2.0",
+    "@spartacus/storefinder": "4.2.1",
+    "@spartacus/smartedit": "4.2.1",
+    "@spartacus/product": "4.2.1",
+    "@spartacus/product-configurator": "4.2.1",
+    "@spartacus/organization": "4.2.1",
+    "@spartacus/checkout": "4.2.1",
+    "@spartacus/cart": "4.2.1",
+    "@spartacus/cds": "4.2.1",
+    "@spartacus/assets": "4.2.0",
+    "@spartacus/asm": "4.2.1",
+    "@spartacus/digital-payments": "4.2.1",
+    "@spartacus/tracking": "4.2.1",
+    "@syncpilot/bpool-guest-lib": "^0.3.2",
     "bootstrap": "4.5.0",
     "echarts": "^5.0.2",
     "ngx-echarts": "^6.0.1",
     "resize-observer-polyfill": "^1.5.1",
     "rxjs": "^6.6.0",
-    "zone.js": "~0.10.3",
+    "zone.js": "~0.11.4",
     "ngx-infinite-scroll": "^8.0.0"
   }
 }
