@@ -10,7 +10,7 @@ import {
 } from '@spartacus/core';
 import { ModalRef, ModalService } from '@spartacus/storefront';
 import { Observable, of, Subscription } from 'rxjs';
-import { filter, map, shareReplay, switchMap, take, tap } from 'rxjs/operators';
+import { filter, map, switchMap, take, tap } from 'rxjs/operators';
 import {
   FSCheckoutConfigService,
   CategoryService,
@@ -79,9 +79,7 @@ export class QuoteReviewComponent implements OnInit, OnDestroy {
 
   isCartTransferAllowedForSeller$: Observable<
     boolean
-  > = this.oboConsentService
-    .isCartTransferAllowedForSeller()
-    .pipe(shareReplay());
+  > = this.oboConsentService.isCartTransferAllowedForSeller();
 
   ngOnInit() {
     this.cart$ = this.cartService.getActive();
@@ -100,11 +98,18 @@ export class QuoteReviewComponent implements OnInit, OnDestroy {
     });
   }
 
-  navigateNext(nextStep: FSSteps, activeCart: Cart) {
+  navigateNext(
+    nextStep: FSSteps,
+    activeCart: Cart,
+    isCartTransferAllowedForSeller: boolean
+  ) {
     this.cartCode = activeCart.code;
     const bindingState = (<FSCart>activeCart).insuranceQuote.state.code;
     const quoteWorkflowState = (<FSCart>activeCart).insuranceQuote
       .quoteWorkflowStatus.code;
+    if (!isCartTransferAllowedForSeller) {
+      return;
+    }
     if (bindingState === BindingStateType.UNBIND) {
       this.openQuoteBindingModal(nextStep);
     } else if (
@@ -171,10 +176,8 @@ export class QuoteReviewComponent implements OnInit, OnDestroy {
   }
 
   getFormContent(cart: any): any {
-    if (cart?.deliveryOrderGroups[0]?.entries[0]?.formData?.length > 0) {
-      return JSON.parse(
-        cart.deliveryOrderGroups[0].entries[0].formData[0].content
-      );
+    if (cart?.entries[0]?.formData?.length > 0) {
+      return JSON.parse(cart.entries[0].formData[0].content);
     }
   }
 
