@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { combineLatest, Observable } from 'rxjs';
-import { map, switchMap, take } from 'rxjs/operators';
+import { filter, map, switchMap, take } from 'rxjs/operators';
 import {
+  PaginationModel,
   RoutingService,
   TranslationService,
   UserIdService,
@@ -9,6 +10,7 @@ import {
 import { ICON_TYPE } from '@spartacus/storefront';
 import { OBOCustomerList, FSUser } from '../../../occ/occ-models/occ.models';
 import { ConsentConnector } from '../../../core/my-account/connectors/consent.connector';
+import { FSSearchConfig } from '../../../core/my-account/services/inbox-data.service';
 
 @Component({
   selector: 'cx-fs-seller-dashboard-list',
@@ -25,11 +27,38 @@ export class SellerDashboardListComponent {
 
   sort: string;
   iconTypes = ICON_TYPE;
+  userId: string;
+  searchConfig: FSSearchConfig = {
+    currentPage: 0,
+  };
 
   customers$: Observable<OBOCustomerList> = this.userIdService.getUserId().pipe(
     take(1),
-    switchMap(userId => this.consentConnector.getOBOCustomerList(userId))
+    switchMap(userId => {
+      this.userId = userId;
+      return this.getOBOCustomerList(userId);
+    })
   );
+
+  getOBOCustomerList(userId: string) {
+    return this.consentConnector
+      .getOBOCustomerList(userId, this.searchConfig)
+      .pipe(
+        filter(customers => !!customers),
+        map(customers => {
+          const pagination: PaginationModel = {
+            currentPage: customers.pagination?.page,
+            pageSize: customers.pagination?.count,
+            totalPages: customers.pagination?.totalPages,
+            totalResults: customers.pagination?.totalCount,
+          };
+          return {
+            ...customers,
+            pagination,
+          };
+        })
+      );
+  }
 
   getUserProfile(customer: FSUser) {
     this.routingService.go({
@@ -56,5 +85,10 @@ export class SellerDashboardListComponent {
         };
       })
     );
+  }
+
+  pageChange(pageNumber: number) {
+    this.searchConfig.currentPage = pageNumber;
+    this.customers$ = this.getOBOCustomerList(this.userId);
   }
 }
