@@ -31,13 +31,15 @@ export class FinalReviewComponent implements OnInit {
     );
   }
 
-  toggleTAndC(): void {
-    this.tAndCToggler = !this.tAndCToggler;
+  toggleTAndC(event: Event): void {
+    this.tAndCToggler = (event.target as HTMLInputElement)?.checked;
   }
 
   placeOrder(): void {
-    this.checkoutService.placeOrder(this.tAndCToggler);
-    this.checkoutService.orderPlaced = true;
-    this.routingService.go({ cxRoute: 'orderConfirmation' });
+    if (this.tAndCToggler) {
+      this.checkoutService.placeOrder(this.tAndCToggler);
+      this.checkoutService.orderPlaced = true;
+      this.routingService.go({ cxRoute: 'orderConfirmation' });
+    }
   }
 }
