@@ -1,4 +1,7 @@
-export const environment = {
-    production: true,
-    occBaseUrl: 'https://10.27.246.242:9012',
+import { Environment } from './models/environment.model';
+
+export const environment: Environment = {
+  production: true,
+  occBaseUrl: 'https://10.27.246.242:9012',
+  digitalPayments: false,
 };
