@@ -1,4 +1,7 @@
-export const environment = {
+import { Environment } from './models/environment.model';
+
+export const environment: Environment = {
   production: true,
   occBaseUrl: 'https://10.27.241.19:9012',
+  digitalPayments: false,
 };
