@@ -1,4 +1,7 @@
-export const environment = {
+import { Environment } from './models/environment.model';
+
+export const environment: Environment = {
   production: false,
   occBaseUrl: 'https://financialservices.local:9002',
+  digitalPayments: false,
 };
